Methods 1, 2, vector_holder, classifier
Implmented methods 1 and two. Compared the output vector of method 1
with the one produced in matlab. They were the same for as much as I
checked.

main_test is a testbed to make sure vector active_nonactive data can be
accessed.

Code is ugly atm...ector holder is probably unnecessary but I made it
when i had a different mindset and the lists of lists thing was acting
odd so I decided to use lists of objects for now.

Fold is still not a thing in python 3

changed the 2nd database file to be 4 rows of 16 items

changed readfile, it was only pulling in 15 items per row instead of 16